Simplify locale dynamic loading
