+++ .babelrc
@@ -14,6 +14,7 @@
     "development": {
         "plugins": [
           "add-module-exports",
+          "dynamic-import-webpack",
           "transform-object-rest-spread",
           "transform-class-properties",
           "transform-react-constant-elements",
