+++ configs/general.js
@@ -11,7 +11,6 @@ export default {
     hashingSalt: '6cee6c6a420e0573d1a4ad8ecb44f2113d010a0c3aadd3c1251b9aa1406ba6a3',
     //Public reCAPTCHA key
     publicRecaptchaKey: '6LdNLyYTAAAAAINDsVZRKG_E3l3Dvpp5sKboR1ET',
-    locales: ['en', 'de', 'it', 'es', 'nl', 'el', 'ca', 'sr'],
     loglevel: 'debug',
     ssoEnabled: true
 };
