+++ general.js.template
@@ -2,7 +2,6 @@ export default {
     fullTitle: ['SlideWiki -- Authoring platform for OpenCourseWare'],
     shortTitle: ['SlideWiki'],
     hashingSalt: '6cee6c6a420e0573d1a4ad8ecb44f2113d010a0c3aadd3c1251b9aa1406ba6a3',
-    locales: ['en', 'de', 'it', 'es', 'nl', 'el', 'ca', 'sr'],
     publicRecaptchaKey: '${SERVICE_USER_PUBLIC_RECAPTCHA_KEY}',
     loglevel: '${LOGGING_LEVEL}',
     ssoEnabled: ${SSO_ENABLED},
