+++ package.json
@@ -167,11 +167,12 @@
   },
   "devDependencies": {
     "babel-eslint": "^7.1.1",
+    "babel-plugin-dynamic-import-webpack": "^1.0.2",
     "chai": "^4.0.2",
     "coveralls": "^3.0.0",
+    "css-loader": "^0.28.7",
     "enzyme": "3.1.0",
     "enzyme-adapter-react-15": "1.0.2",
-    "css-loader": "^0.28.7",
     "eslint": "^3.17.0",
     "eslint-plugin-babel": "^4.1.1",
     "eslint-plugin-react": "^6.10.0",
@@ -183,6 +184,7 @@
     "nodemon": "1.11.0",
     "react-addons-test-utils": "^15.6.0",
     "react-hot-loader": "1.3.1",
+    "react-intl-loader": "^1.0.2",
     "react-test-renderer": "15.6.2",
     "shelljs": "^0.7.8",
     "sloc": "0.2.0",
