--- server/intl-polyfill.js
@@ -1,11 +1,0 @@
-/* global Intl */
-
-// Add support for intl on node.js
-// See: http://formatjs.io/guides/runtime-environments/#server
-
-import { locales } from '../config/general';
-import areIntlLocalesSupported from 'intl-locales-supported';
-
-if (!global.Intl || !areIntlLocalesSupported(locales)) {
-    require('intl');
-}
